refactor: clean up settings exports
