--- renderer/components/Settings/Settings.js
@@ -1,3 +1,0 @@
-import React from 'react'
-
-export const SettingsContext = React.createContext({})
